calling-agent/requirements.txt
flask==3.0.0
google-cloud-speech==2.26.0
google-cloud-pubsub==2.21.0
google-cloud-aiplatform==1.50.0
